move strset to a new package
