+++ Dockerfile
@@ -15,6 +15,7 @@ COPY api/ api/
 COPY clusterversion/ clusterversion/
 COPY collector/ collector/
 COPY controllers/ controllers/
+COPY strset/ strset/
 
 # Build
 RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 GO111MODULE=on go build -a -o manager main.go
