+++ collector/types.go
@@ -291,28 +291,3 @@ func (row StorageRow) RowString() []string {
 func (row StorageRow) String() string {
 	return strings.Join(row.RowString(), ",")
 }
-
-var exists = struct{}{}
-
-type set struct {
-	m map[string]struct{}
-}
-
-func NewSet() *set {
-	s := &set{}
-	s.m = make(map[string]struct{})
-	return s
-}
-
-func (s *set) Add(value string) {
-	s.m[value] = exists
-}
-
-func (s *set) Remove(value string) {
-	delete(s.m, value)
-}
-
-func (s *set) Contains(value string) bool {
-	_, c := s.m[value]
-	return c
-}
