+++ strset/strset.go
@@ -1,0 +1,28 @@
+package strset
+
+var exists = struct{}{}
+
+// Set is simple set for strings
+type Set struct {
+	m map[string]struct{}
+}
+
+// NewSet creates a new *set
+func NewSet() *set {
+	s := &set{}
+	s.m = make(map[string]struct{})
+	return s
+}
+
+func (s *set) Add(value string) {
+	s.m[value] = exists
+}
+
+func (s *set) Remove(value string) {
+	delete(s.m, value)
+}
+
+func (s *set) Contains(value string) bool {
+	_, c := s.m[value]
+	return c
+}
